+++ trunk/TeleAsistenciaPasos/TeleAsistenciaPasos-ejb/src/java/org/inftel/pasos/ejb/UsuarioFacade.java
@@ -4,7 +4,6 @@
  */
 package org.inftel.pasos.ejb;
 
-import java.math.BigDecimal;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -28,6 +27,6 @@ public class UsuarioFacade extends AbstractFacade<Usuario> {
     }
     public Usuario findById(Integer id) {
         return (Usuario) em.createQuery("SELECT u FROM Usuario u WHERE u.idUsuario = :id").
-                setParameter("id", new BigDecimal(id)).getResultList().get(0);
+                setParameter("id", new Long(id)).getResultList().get(0);
     }
 }
